requirements.txt
# Core dependencies
pyyaml>=6.0
watchdog>=3.0.0
httpx>=0.24.0
websocket-client>=1.6.0
requests>=2.31.0

# Web framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
jinja2>=3.1.0

# AI/LLM dependencies
langchain>=0.0.350
langchain-openai>=0.0.5

# Image processing
Pillow>=10.0.0

# Database (using stdlib sqlite3, but adding sqlalchemy for ORM if needed)
# sqlalchemy>=2.0.0  # Optional - uncomment if using ORM

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0

# Code quality
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# CLI
typer>=0.9.0
rich>=13.5.0  # For better CLI output

# Utilities
python-dotenv>=1.0.0  # For .env file support
beautifulsoup4>=4.12.0  # For HTML parsing
pydub>=0.25.0  # For audio processing and combining

# Additional dependencies found in codebase
glob2>=0.7  # For advanced file globbing (if needed)
subprocess32>=3.5.4  # Better subprocess handling (Python 2 backport, optional for Python 3)

# Development and debugging
ipython>=8.0.0  # For interactive development